finish stats, reorganize modules
