+++ __tests__/processors.js
@@ -1,8 +1,8 @@
 jest.autoMockOff();
 
 const _ = require('underscore');
-const DDPGenerator = require('../src/ddp-generator');
-const TraceProcessor = require('../src/trace-processor');
+const DDPGenerator = require('../src/lib/ddp-generator');
+const TraceProcessor = require('../src/lib/trace-processor');
 
 let runProcessor = (traces, isOutbound=true) => {
   return TraceProcessor.processTraces(
