+++ src/components/stats.jsx
@@ -1,5 +1,4 @@
 import React, { PropTypes } from 'react';
-
 export default React.createClass({
   propTypes : {
     stats: React.PropTypes.shape({
@@ -8,20 +7,16 @@ export default React.createClass({
       messageTypes: React.PropTypes.object.isRequired
     }).isRequired
   },
-  // onClick (event){
-  //   this.props.onToggle(this.props.name);
-  // },
   render (){
-    // let filterClass = classNames('filter', (this.props.name).toLowerCase());
-    // let filterId = 'hide-'+(this.props.name).toLowerCase();
-
     return (
       <div>
-        <strong>in: {this.props.stats.inboundMessages}</strong>&nbsp;
-        <strong>out: {this.props.stats.outboundMessages}</strong>&nbsp;
-        <strong>collections: {this.props.stats.messageTypes.Collections}</strong>&nbsp;
-        <strong>methods: {this.props.stats.messageTypes.Methods}</strong>&nbsp;
-        <strong>subscriptions: {this.props.stats.messageTypes.Subscriptions}</strong>
+        <i className="fa fa-arrow-down"></i>&nbsp;
+        <strong>{this.props.stats.inboundMessages}</strong>&nbsp;
+        <i className="fa fa-arrow-up"></i>&nbsp;
+        <strong>{this.props.stats.outboundMessages}</strong>&nbsp;|&nbsp;
+        <strong>collection ops: {this.props.stats.messageTypes.collections}</strong>&nbsp;
+        <strong>method calls: {this.props.stats.messageTypes.methodCalls}</strong>&nbsp;
+        <strong>subs: {this.props.stats.messageTypes.subscriptions}</strong>
       </div>
     )
   }
