+++ src/components/trace-item-prop-types.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 
 export default {
   data: React.PropTypes.shape({
+    _id : React.PropTypes.string.isRequired,
     isOutbound : React.PropTypes.bool.isRequired,
     stackTrace : React.PropTypes.array,
     message : React.PropTypes.object.isRequired,
