+++ src/components/trace-list.jsx
@@ -33,7 +33,7 @@ export default React.createClass({
     const noData = this.props.traces.length === 0 ?
       <li className="no-trace">No traces yet...</li> : null; 
     const items = this.props.traces.map(function(item){
-      return <TraceItem data={item}/>;
+      return <TraceItem data={item} key={item._id}/>;
     });
 
     return (
