+++ src/lib/processors/operation-types.js
@@ -1,9 +1,9 @@
 import _ from 'underscore';
-import operationTypes from '../constants/operation-types'
+import operationTypes from '../../constants/operation-types'
 
 class TraceOperationTypes {
   run(traces) {
-    const results = _.map(traces, (t) => {
+    return _.map(traces, (t) => {
       const operationType = _.find(_.keys(operationTypes), (ot) => {
         return _.contains(operationTypes[ot], t.operation);
       });
@@ -11,10 +11,6 @@ class TraceOperationTypes {
         operationType : operationType
       });
     });
-
-    console.error('traces with op types', results);
-
-    return results;
   }
 };
