+++ src/lib/trace-filter.js
