+++ src/lib/trace-processor.js
